// dist/node/community/schema.d.ts
export declare const ChallengeFileSchema: z.ZodObject<{
    optionInputs: z.ZodOptional<z.ZodArray<z.ZodObject<{
        option: z.ZodString;
        label: z.ZodString;
        default: z.ZodOptional<z.ZodString>;
        description: z.ZodOptional<z.ZodString>;
        placeholder: z.ZodOptional<z.ZodString>;
        required: z.ZodOptional<z.ZodBoolean>;
    }, z.core.$loose>>>;
    type: z.ZodString;
    challenge: z.ZodOptional<z.ZodString>;
    caseInsensitive: z.ZodOptional<z.ZodBoolean>;
    description: z.ZodOptional<z.ZodString>;
    getChallenge: z.ZodFunction<z.ZodTuple<readonly [z.ZodObject<{
        challengeSettings: z.ZodObject<{
            path: z.ZodOptional<z.ZodString>;
            name: z.ZodOptional<z.ZodString>;
            options: z.ZodOptional<z.ZodRecord<z.ZodString, z.ZodString>>;
            exclude: z.ZodOptional<z.ZodArray<z.ZodObject<{
                community: z.ZodOptional<z.ZodObject<{
                    addresses: z.ZodArray<z.ZodString>;
                    maxCommentCids: z.ZodNumber;
                    postScore: z.ZodOptional<z.ZodNumber>;
                    replyScore: z.ZodOptional<z.ZodNumber>;
                    firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
                }, z.core.$strict>>;
                postScore: z.ZodOptional<z.ZodNumber>;
                replyScore: z.ZodOptional<z.ZodNumber>;
                postCount: z.ZodOptional<z.ZodNumber>;
                replyCount: z.ZodOptional<z.ZodNumber>;
                firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
                challenges: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
                role: z.ZodOptional<z.ZodArray<z.ZodUnion<[z.ZodEnum<{
                    owner: "owner";
                    admin: "admin";
                    moderator: "moderator";
                }>, z.ZodString]>>>;
                address: z.ZodOptional<z.ZodArray<z.ZodString>>;
                rateLimit: z.ZodOptional<z.ZodNumber>;
                rateLimitChallengeSuccess: z.ZodOptional<z.ZodBoolean>;
                publicationType: z.ZodOptional<z.ZodObject<{
                    post: z.ZodOptional<z.ZodBoolean>;
                    reply: z.ZodOptional<z.ZodBoolean>;
                    vote: z.ZodOptional<z.ZodBoolean>;
                    commentEdit: z.ZodOptional<z.ZodBoolean>;
                    commentModeration: z.ZodOptional<z.ZodBoolean>;
                    communityEdit: z.ZodOptional<z.ZodBoolean>;
                }, z.core.$loose>>;
            }, z.core.$loose>>>;
            description: z.ZodOptional<z.ZodString>;
            pendingApproval: z.ZodOptional<z.ZodBoolean>;
        }, z.core.$strict>;
        challengeRequestMessage: z.ZodCustom<DecryptedChallengeRequestMessageTypeWithCommunityAuthor, DecryptedChallengeRequestMessageTypeWithCommunityAuthor>;
        challengeIndex: z.ZodNumber;
        community: z.ZodCustom<LocalCommunity, LocalCommunity>;
    }, z.core.$strip>], null>, z.ZodPromise<z.ZodUnion<[z.ZodObject<{
        challenge: z.ZodString;
        verify: z.ZodFunction<z.ZodTuple<readonly [z.ZodLazy<z.ZodString>], null>, z.ZodPromise<z.ZodUnion<[z.ZodObject<{
            success: z.ZodLiteral<true>;
        }, z.core.$strip>, z.ZodObject<{
            success: z.ZodLiteral<false>;
            error: z.ZodString;
        }, z.core.$strip>]>>>;
        type: z.ZodString;
        caseInsensitive: z.ZodOptional<z.ZodBoolean>;
    }, z.core.$strict>, z.ZodUnion<[z.ZodObject<{
        success: z.ZodLiteral<true>;
    }, z.core.$strip>, z.ZodObject<{
        success: z.ZodLiteral<false>;
        error: z.ZodString;
    }, z.core.$strip>]>]>>>;
}, z.core.$strict>;